Fixed bug that causes dialer.exe to fail:
there is a deadlock between tapi provider and websock service: when websockservice wants to sent a message to the tapi provider, it opens a line for that. After it sends message, it closes the line. The line close handling in tapi provider implies disconnect from the service. That leads to the deadlock.

I put the disconnect into a separate thread.